experimental/agno_version/requirements.txt
# ========================================
#   agno-mirofish 依赖包
#   在原 BettaFish requirements.txt 基础上新增 agno
# ========================================

# ===== agno 框架（新增）=====
agno>=1.7.0

# ===== 核心Web框架 =====
flask==2.3.3
flask-socketio==5.3.6
streamlit==1.28.1
python-socketio==5.8.0
eventlet==0.33.3

# ===== HTTP请求和异步 =====
requests==2.31.0
httpx==0.28.1
aiofiles==23.2.1
aiohttp>=3.8.0

# ===== LLM接口 =====
openai>=1.3.0

# ===== 搜索API =====
tavily-python>=0.3.0

# ===== 数据处理 =====
pandas>=2.0.0
numpy>=1.24.0
regex>=2023.8.8
jieba==0.42.1

# ===== 数据库 =====
pymysql==1.1.0
SQLAlchemy==2.0.35
asyncpg==0.29.0
psycopg[binary]>=3.1.0

# ===== 爬虫相关 =====
playwright==1.45.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# ===== 可视化 =====
plotly>=5.17.0
matplotlib==3.9.0
wordcloud==1.9.3

# ===== PDF生成 =====
weasyprint>=60.0

# ===== 机器学习 =====
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.2
scikit-learn>=1.3.0

# ===== 工具库 =====
python-dotenv>=1.0.0
tenacity==8.2.2
loguru>=0.7.0
pydantic==2.5.2
pydantic-settings==2.2.1
json-repair==0.53.0
tqdm>=4.65.0

# ===== 开发工具 =====
pytest>=7.4.0

# ===== Web服务器 =====
fastapi==0.110.2
uvicorn==0.29.0
